reader iterators unification
